+++ iris/__main__.py
@@ -17,7 +17,7 @@ import pytesseract
 import firefox.app as fa
 import firefox.downloader as fd
 import firefox.extractor as fe
-from api.core_helper import get_module_dir, get_platform, get_run_id, get_image_debug_path
+from api.helpers.core_helper import get_module_dir, get_platform, get_run_id, get_image_debug_path
 from api.helpers.parse_args import parse_args
 from api.keys import Key
 from api.platform_iris import Platform
