+++ iris/api/location.py
@@ -2,7 +2,7 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this file,
 # You can obtain one at http://mozilla.org/MPL/2.0/.
 
-from core_helper import INVALID_NUMERIC_INPUT
+from helpers.core_helper import INVALID_NUMERIC_INPUT
 
 
 class Location(object):
